Updating index page tests
